add handler for fpe, bus error
