+++ python/mxnet/error.py
@@ -57,3 +57,7 @@ register_error("TypeError", TypeError)
 register_error("AttributeError", AttributeError)
 register_error("IndexError", IndexError)
 register_error("NotImplementedError", NotImplementedError)
+register_error("InternalError", InternalError)
+register_error("IOError", IOError)
+register_error("FloatingPointError", FloatingPointError)
+register_error("RuntimeError", RuntimeError)
